Fix nits.
