+++ src/main/webapp/mentor_breakout.html
@@ -25,5 +25,5 @@
   
   <script src="https://cdn.jsdelivr.net/npm/mailtoui@1.0.3/dist/mailtoui-min.js"></script>
 
-  <script type="module" src="mentors.js"></script>
+  <script type="module" src="mentor_breakout.js"></script>
 </body>
